requirements.txt
langchain
chromadb
pypdf
sentence-transformers
groq
langchain_community
python-dotenv
langchain_groq
beautifulsoup4
requests